app/requirements.txt
streamlit>=1.32
pandas>=2.0
plotly>=5.18
reportlab>=4.1
python-dateutil>=2.8
